Adjust potion of mutation probabilities
It was very easy to reliably remove bad mutations with just a few potions
of mutation due to them removing 2-4 mutations as their first effect, and
to acquire a mostly or entirely positive mutation set due to the guaranteed
good mutation at the end.

While they should still be useful for clearing out bad mutations, the
consistency was a little too high, so reduce the number of mutations
removed to 2-3, and reduce the chance of adding a good mutation at the end
from 100% to 50% (so that the overall number of mutations added/removed on
average also remains the same).