Remove wispform's blinkitis and extra speed
Random blinking was an especially annoying feature even for a form that's
intended to be dangerous. The monster that the form was based on also no
longer exists, so instead just make wispform standard speed (distinct from
bat/pig badforms which are fast), keeping the existing downsides of most
other badforms of melding equipment and preventing weapon or spell use.