Unrename a vault
This vault was apparently intentionally renamed in 2394d0fb to change
"ooze" to "endoplasm", but the vault never placed a monster that got
renamed to endoplasm. It did place a death ooze, which got replaced with
a quicksilver ooze, so the original name is still appropriate. In any
case, renames should be avoided unless the old name would somehow
generate confusion or a new naming format is used for a set of related
vaults, neither of which is the case here.