feat: Rework a monster float vault (again)
This commit finishes aspects of gammafunk_indiana_jones I wasn't able to
do in the last commit. It reworks the final chamber to feature lindwurms
safely protected from fire clouds while still having lava snakes,
putting the fire cloud machines on the lava directly. It also adds a
Tomb theme for D:14- and Spider featuring death scarabs and spectrals in
the final chamber. Loot, monster, and trap  placement progression with
placement depth is now consistent and should pose a nicely scaling
challenge and reward for all levels. Other slight tweaks have been made
to the layout to more consistently apply the visual theme, including
statues and plant "overgrowth" in the starting area, as well as statues
in the final chamber. All doors inside are runed doors, which is the
best balance between UI and preventing spoilers.