feat: add a way to reset the !a warning without swapping (Yermak)
This is a very minimal change to allow this, it resets the warning as a
side effect of rewielding. This is not great UI and it's a bit of a
hidden feature, but as Yermak pointed out, there are corner cases where
you really want something like this (e.g. Ash, + accidentally hitting y
or using it one-off for something like rampaging). I've at least
documented the behavior. This flag is used for ranged weapons as well as
!a-inscribed weapons.