Adjustments to two volcano maps
For hellmonk_zot_springs, thin out the upper-tier monsters in the main
chamber, since it felt a bit too fierce in play-testing. The middle-tier
molten gargoyles are pretty tough at this depth as it is, so multiple
salamander mystics and the other higher tier threats compound the
difficulty of this map with each additional placement. Make the
upper-tier set not include molten gargoyles, so we won't place too many
of those, since the middle-tier placement already places four
on-average.

For syntax, reglyph to use consistent glyphs for monsters and items and
to use a linear glyph progression. Fix the middle-tier monster NSUBST of
3 to not use ':' when placing the next tier on 4, since this would place
either six of such monsters or zero, which seems unintentional (and
undesirable). Clean up water tile placement and reorganize statements.

For hellmonk_pyromancer_palace, reglyph decorative lava so we can mark
it no_tele_into without marking the non-decorative lava.

Closes #1949.