feat: rescale apts from [-5,+5] to [0,10]
At the moment, this commit is intended to be up for discussion.

On the one hand, the positive/negative values for aptitudes provide an
easy and convenient shorthand for "good"/"bad", and also make it
relatively easy to do the math for what a particular apt will mean. On
the other hand, we have seen that this scheme often leads to players
overemphasizing the sign vs the actual numeric value; no amount of
documentation etc has led to players (especially newer ones) conceiving
of -1 is more "normal" as opposed to "terrible". For reference (sorry,
did not recalculate this for 0.27, but it hasn't changed that much): in
0.25, the overall mean aptitude was -0.35, with the worst species being
troll at -2.6 and the best species being barachi at +0.77.

This commit is a way I've found to solve some aspects of this problem
without majorly changing the underlying math: where previously
everything except Gn fell on a -5 to +5 range, it simply linearly
rescales aptitudes onto a 0 to 10 range for presentational purposes.
(Internally, nothing changes.)

I do think this makes the skill menu a bit harder to process quickly,
and I've accordingly done some recoloring to try to make good and bad
apts stand out more quickly. But, this still may be a net negative in UI
terms.