dreamdust_venomous_pit & gu_poison
venomous_pit - A very small taste of the Snake Pit in the early-game.
gu_poison - Based on Chinese folklore about creating deadly (often magical) poisons by putting a bunch of poisonous/venomous bugs in a sealed jar and having them fight until only one is left alive. The winner was thought to have absorbed and concentrated the poisons of the losers.