Fix spell -Potion application (AliasTheSpectator)
If the player had rC+++ and cast Ozocubu's Refrigeration, their immunity
to its damage accidentally prevented application of -Potion. The player
spell doesn't damage the player even without the immunity granted by
rC+++, but this bug emerged as a holdover from when the spell did
self-damage. This commit applies the debuff unconditionally.