More cleanups for Ozocubu's Refrigeration
The check for whether a target was vulnerable used some outdated logic,
and the UI for players casting this spell didn't even perform the check
in the first place. Hence players casting the spell with only rC+++
monsters in LOS would not get the usually prompt about a useless cast,
even though all visible targets are immune. This commit fixes this
logic, adds a tracer check, and refactors the LOS attack spell code
slightly.