Lower Absolute Zero's range to 3
This spell has a penalty of randomly selecting the target, but because of its long range, the player was able to offset the penalty by simply using the spell several times. Thus, the Pandemonium lord turned into an ice pillar upon encountering the player. So, to curb brainless play, I'm trying to reduce the range of this spell to 3.