Reduce monster count in a vault (Ge0ff)
The vault nzn_swamp_stair_ambush places 8 guaranteed monsters near
stairs in a small area. While this is within acceptable bounds, it's on
the hard side of things, and additionally there's no monster count
randomization, which we like to have for monster-placing vaults. Remove
three monsters on-average by adding appropriate chances for shallow
water or floor instead of a monster in some places.