Abyssal woods Lair end
Adds a new lair end vault based on enchanted_forest, but makes use of demonic
trees and an abyssal monster set. Thrashing horrors are the big threat here,
so adjusting the number of those would be the easiest way to modify difficulty.

Committer's Note: The subsequent commit reworks parts of this map.