Clean up lua depth checks
Use you.in_branch() instead of a comparison to the you.depth() string in
order to check branch location, since the latter is error-prone. This
batch of changes only caught one such error, but the string comparison
is case-sensitive, and you.in_branch() correctly generates an error if
the argument given is an invalid branch name.

Also prefer use of you.depth() to you.absdepth() when only one branch is
involved, as this is clearer. It would probably be best to just always
use a combination of you.in_branch() and you.depth(), since its more
clear and doesn't require DES readers to look up branch absdepth values.
However this requires rewriting a lot of conditionals, and I've mostly
only modified one related to the first part of this patch.