Simplifiy Invisibility (Yermak)
Make it self-targeted only, to reduce the number of keystrokes while
casting. It was possible to use Invisibility on other targets, e.g.
on allies as a buff or on enemies to trick enemies into shooting them,
but in practice this happened fairly rarely. In future, we might add
some other effect (perhaps an evocable or unrand) that encourages those
fun interactions.