Remove Okawaru's friend conduct
It had little actual impact on gameplay and little flavour - mostly it
was an occasional surprising annoyance. I think there's room here for
a much more interesting, stronger anti-ally conduct, but I don't know
when we'll get to it, and there's no good reason to leave this around
in the meanwhile.