Update IRC and community information
Now that we've completed the move to Libera IRC from Freenode, update
all references to that network and the appropriate channels. I took the
liberty of updating two very minor source comments in cc/DES files to
refer to conversations as being on "IRC" instead of "##crawl", so anyone
reading them without pior knowledge of the old channel won't be
confused.

I've also updated the Overview and section of the manual to refer
primarily to the community resource listing on http://crawl.develz.org
instead of just listing IRC. I've added a "Contact and community
information" heading to the "License, contact, history" section and put
the existing community info under this heading. I've also added the DCSS
subreddit to the list of community resources and adjusted the language
to be applicable to both it and Discord. I've updated the IRC info to
mention #crawl-dev for development. History is moved to its its own
heading and will be updated in a subsequent commit.

For Hints and FAQs, again refer to the website for community resources
instead of just listing IRC, and update IRC listing for hints in other
languages. Ideally these hints would instead refer to the website, but
that would need work by translators. I've also referenced the github
issue tracker over mantis in a couple places where only mantis was
mentioned and referring to both would be awkward. Possibly we need to
consistently mention only the github tracker and only put information
about mantis in somewhere like the contributor's guide.