Shorten Gozag's distraction message
It was a lot wordier than necessary for something that's often displayed
multiple times a turn. The "dreaming of imaginary riches" part was intended
to clarify that it doesn't require actual gold to be present on the floor
or in inventory, but the new "golden aura" wording now covers that anyway.