Move Freezing Cloud to level 5
Since clouds now dissipate out of LOS and Ozocubu's Refrigeration
now deals heavy damae, it's hard to justify the investment in Freezing
Cloud as a 3-school spell. Move it to level 5 to make it a bit more
viable, and to add symmetry with Bolt of Magma as a L5 Fire/Earth/Cj.