fix: let `sword_in_stone` not choke on dancing singing sword
e7febcc0c011a added singing sword to `sword_in_stone`. However, the
dancing weapon spec parsing couldn't handle providing an unrand here,
and so the 1/50 condition would just produce a lua error (even if the
vault is placing at a depth where the non-dancing one would be used).
This commit implements correct spec parsing for unrands, and so lets
this vault work as intended. If singing sword is already placed, it
places a dancing fallback unrand (which seems to work correctly).