fix: various vault transparency fixes
Mostly for new-ish vaults; I'm a bit surprised if these placed. In most
cases this is just adjusting the masking. For nemelex_altar_shiori the
removal of the transparent tag may be awkward, as it means this will
veto unless *both* doors link up, but masking this one would be a pain.