Shift: Tornado (L9 Air) -> Polar Vortex (L9 Ice)
Ice, as a school, is intended to be 'diffuse' and 'subtle', With a
few exceptions (i.e. Freeze), it tends to be best at dealing damage
gradually to many enemies. That provides it with a distinctive
playstyle, and Tornado fits that playstyle perfectly.

As such, we'll shift Tornado to be the new level 9 Ice spell.
Subsequent commits will discuss the removal of Absolute Zero and the
new level 9 Air spell further.

Functionally, the main change here is that Polar Vortex does 33% more
damage (3 more dice) but is 40% resistable by rC, as with Hailstorm.
Otherwise, it's the same spell.