Remove Absolute Zero
Absolute Zero was a spectacular, memorable effect: instantly turn the
closest enemy into a block of ice! Unfortunately, in play, it ended up
feeling worse than it sounded. The targeting constraints were a bit less
limiting than had been intended, and many otherwise dramatic fights
against powerful foes felt like anticlimaxes. It's still not entirely
clear whether the spell was "too powerful" relative to the investment
required to cast it, but it didn't *feel* good.

There have been many suggestions for improving Absolute Zero - for example,
adding a delay to the effect and making it a L9 Poison (!!) spell. When
we want to experiment with those, we can revert this commit. For now,
though, let's toss it.

(A small part of this change [targeter removal] is in the previous
commit - sorry! Don't miss that if you revert this.)