Some Shoals vaults from nicolae (12593)
A set of 7 decor vaults and a set of 18 monster vaults. Minor edits have
been made to the original submissions to follow syntax ordering style
and to reglyph monsters to use number glyphs where possible. Remove
patrolling tag for vaults that use in-branch monsters and add it in
other cases where out-of-branch monsters are placed in vaults with
special themes. More substantive edits to a couple vaults described
below:

nicolae_shoals_hidden_pirate_stash: Make the possible rapier weapon loot
be good_item and make the scroll 50/50 for magic mapping or any scroll,
instead of having a weight of 4 for the latter.

nicolae_shoals_pincushion_passage: Thin out the monster a count, turning
a 2/3 chance into a 1/3 chance, since otherwise it's a bit high for
placing so many upper tier monsters. Could go either way on this change,
but a bit fewer monsters felt right.

nicolae_shoals_raging_river: Use elemental wellsprings instead of a
custom vault define, which was a permanently berserk water elemental. A
cute idea, but ultimately not quite interesting enough to warrant a
custom definition, and I think what wellsprings can do is ultimately
more interesting.