fix: prevent save corruption on &P crashes
This fixes a couple of ways that a save could become unloadable when it
crashes with &P: (i) the force props didn't get cleared, so the game
would try to regenerate exactly the failing levelgen process all over
again (which might be ok for some crashes, but not e.g. a bad itemspec),
and (ii) placeinfo would get corrupted in a way that doesn't get fixed
up (now it gets corrupted in a way that does).