fix: improve crashes for unrecoverable builder lua errors
This isn't intended to be a full solution for lua errors at levelgen time, but
tries to make the crash process smoother, both in terms of messaging,
and efficiency.

The builder for each map tries to run vault lua 25 times, and tries to
build a level 50 times before bailing entirely with a crash. So if a
vault has e.g. a broken itemspec that produces a lua error during
levelgen, this can be repeated at least 25 times, and for wizmode force
generation (&P), 25 * 50 times. The vast majority of errors of this
category (though not all) will be deterministic, so there's no point to
retries. In cases like this, on a debug build the lag for the crash is
also pretty noticeable. So, implement some stuff to immediately bail
(for &P) or veto (for regular levelgen). In addition, percolate the
actual lua error up the messaging chain a bit better, so the it now is
actuall visible directly from the command line. (It's probably a bit
*too* redundantly printed now, though.)

Implementation: this commit adds a state variable to signal that an
error was fatal, and handling for it to improve the messaging and skip
all these retries.  This may be something that would be more naturally
done with exceptions, but it's entirely possible that these exceptions
would need to dance around the lua layer (which can't have exceptions
thrown across it, see 091e5a1), and I don't want to mess with this. For
regular levelgen, this does throw a veto exception, but from a place
where that should be safe to do.