feat: tweak polar vortex glyph colours
It's possible this is overcomplicating things, as it may make it
slightly harder to identify terrain. However, the glyph colours were
inconsistent with the tile colours, and also not very on-theme for how
ice is usually depicted on console. (Of course, if Tornado had never
existed, this might be less of an issue.) I ended up deciding that e.g.
just using glaciate glyph colours didn't really work, it was too ugly,
so this uses a bit more randomness for one of the phases to add some
more subtle console ice colouring in. One phase is always stable, to
prevent complete chaos.