fix: remove some ancient mac debugging code
This code originates from pre-version-controlled dcss or earlier. I can
only imagine that at some point <2005 things were weird with intentional
crash signals in MacOS to the point where accessing *nullptr seemed like
a good idea, but on a modern mac this has been increasingly unreliable,
and on an M1 / big sur device, it freezes with 100% cpu before timing
out or needing kill -9, at least 2/3 of the time. As far as I can tell,
sending abort works fine and doesn't do this. We'll see if this
generalizes to other devices.