fix: guaranteed source of digging in sprint_mu slime (Nivim)
There is no code to remove the walls on TRJ death, and it is not easy to
port the regular game code over. I'm guessing the intent was to break
these walls somehow (they are rock, not stone). The sprint does place a
source of digging, just not in slime, possibly because it would have
been eaten.  However, that is no longer an issue, so I'm just adding a
guaranteed digging wand on the altar. If someone ever implements wall
destruction here, remove this.