Rework Poisonous Vapours not to create clouds
Mechanically, Poisonous Vapours is intended to apply a small amount of
smite-targeted poison, but the fact that it was implemented by using a
cloud ended up making it unnecessarily complicated. Almost all of the time
the cloud would dissipate immediately, but at particularly high power it
would stick around for a brief while, and using a cloud also prevented
using it against targets that were already in clouds.

Instead make it just directly apply poison, with the number of stacks
applied scaling up at high power (to compensate for the fact that the cloud
itself could previously deal some direct damage).