Refactor and expand objstat statistics
This commit reworks most of the internal data handling of objstat to be
more consistent and to allow tracking of additional data. All item
categories now track additional categories of whether the item was in a
vault, held by a monster, or in a shop. Aspects like enchant/charges and
brands are likewise tracked within these new categories. We also now
track statistics for spells, as described further below. Artefact books
are moved over to the spellbook tracking file, treating level artefact
books and theme artefact books as two distinct book types. Additionally
manuals now track by their associated skill instead of all together.
Finally, objstat no longer produces redundant listings for "All Levels"
when only one branch is tabulated, since that data is already in the
branch summary.

For items, the tracking categories of artefact, vault, shop, and held by
monster are independent of each other. So one can't determine the number
of items both held by a monster and in a vault, for example; tracking
all combinations of these aspects would lead to too many statistics of
marginal interest. Exception to this are the fact that an item can't be
both held by a monster and in a shop and that all shop items are vault
items, since shops place through vaults. Subtracting the values in these
categories from the overall value will always give the count for the
negation of that category. For example, subtracting NumArte from Num
gives the number of non-artefact items. I've removed fields like
"NumOrd" that can be calculated in this way to reduce the number of
output statistics; these calculations can be easily made by a
spreadsheet, once the data is imported.

For spell tracking, we track overall counts by spell and within the new
categories mentioned above, including artefact for spells in artefact
spellbooks. Additionally we have "Chance" fields for spells in all
categories. The Chance field gives the percentage of the time the spell
was seen on a given level. This can be useful for determining the chance
a spell would be seen by some depth, for example.  Calculation of Chance
fields could be extended to other item categories like Misc, Staves, or
even all items, if that's later deemed useful.

Internally it now uses map<> instead of vector<> for indexing
statistics, which simplifies statistical initialization, since we can
rely on default initialization of zero for most statistics. I've made
these maps use enum types keys where that's possible, but in some cases
like item sub_type we still need to use int keys, since item sub_types
don't share a common enum type.