fix: fully inactivate faith for ashenzari (u/Imaginationably)
c9f47ef48ce0 disabled the piety gain effect of faith and marked it
useless, but did not turn of the on-remove effect or messaging.
(Possibly this code would be cleaner if all of these things were tied to
uselessness, but I'm not sure enough how the other gods work to change
that.) It's still possible to wear useless items (and in this case, even
curse them -- in fact this maybe means that ash followers should collect
useless amulets until they have something better to curse? Haven't
played enough newash to know if this is a good idea.)