fix: don't mark ash-cursed items as useless
In various cases the player can put on items whose base ego is useless
(e.g. faith), and then curse them, in which case they are no longer
useless. The melding logic here relies on the rest of the code being
right, which it probably isn't, but it looked ok in the cases I tested.