Increase book generation rates
Move them from weight 30 to 100, slightly more than doubling the number
of books generated. This is in preparation for a future change that will
shrink books. All other item generation drops by about 2%, which I'm not
currently compensating.  This keep total spells more in line with what
we saw before.