Split fixed books into smaller sizes
To smooth out the vagaries of randomness and make it more viable
for players to adapt to what books they find, shrink most fixed books
to contain exactly 3 spells and add about 30 new books to compensate.

These books are designed to be relevant to a wide variety of character
levels and types, to increase the chance that they'll have at least
one spell that the finder will be able to use, even if it's less likely
that every spell in there will be useful. They're also supposed to have
fun, evocative themes.

TODO: book descriptions (or removal of book descriptions) and more
thorough vault changes (adding to alphashops & themed vaults, fixing
up overly large randbooks).