Refactor Kiku's Necronomicon gift, re-add Haunt
With Haunt being moved out of the Necronomicon it had become impossible for
Kiku to gift knowledge of it. Instead of gifting a physical Necronomicon,
make Kiku add its spells directly to the library as with the earlier spell
gifts, and add Haunt back to the list of included spells.