Mostly remove resilience
This basically just confused a lot of people over the years. It's OK,
you can say 'health' when you mean 'health'. We already use it in a
few places! It's more flavourful for stuff like Borg's anyway - very
Raistlin Majere.

Skeletons and zombies are a bit beyond the whole 'health' thing, and
I think resilience is less confusing for them anyway, so they can keep
it. Ditto the reference on the amulet of vitality - that one's very
straightforward.