feat: better integrate the macro menu
Before, there was a pre-menu to select the kind of binding to edit
before entering the edit menu. After this change you are dropped
directly into macro edit mode, and can use `!` to switch to keybind
editing. This is one fewer keystroke for macros, the same number of
keystrokes for regular keybinds, and extra keystrokes for the obscure
keybind modes. I've also slightly hidden the details of keybinds from
the first view, because in almost all circumstances I don't think
players want to edit these in-game, and in almost all circumstances,
macros are a better choice than keybinds for people who don't know what
they are doing.

Still lots of further improvements to be made here. The editing itself
should be moved onto the same screen as the menu, and I think this menu
would also be more convenient if arrows selected (but unfortunately the
Menu class doesn't support this at all, as far as I know).