fix: set last_visible correctly in webtiles menus
The for-loop here for setting last visible only checks the case where
i-1 is the last visible element, so when the last element is visible, it
didn't do anything, leaving last_visible set to 0. Instead, make the
fallthrough case already the last element in the list. As far as I can
tell this should work even on very strange bounding boxes. As a
practical matter, this will prevent menus from wrapping to the beginning
on pgdn in webtiles, which they apparently have been doing for a while.
We'll see how used to this players are. Other clients have a MF_WRAP
setting for when menus are supposed to have this behavior, but this
doesn't seem to be implemented in webtiles.

This commit also fixes a bunch of js exceptions on corner cases with
empty menus.