feat: arrow key navigation for macros menu
More incremental improvements for the macro menu. The immediate goal of
this is to allow more easily selecting existing macros for editing that
have hotkeys that menu code would otherwise intercept.

Despite the limited application of this change, the commit is a fairly
comprehensive rewrite of hover code that is tested to work on any
single-selection menu. I may activate it more generally in the future as
I think it's a net usability improvement in all sorts of menus (though I
expect a certain amount of grumbling because it is *change*). However,
it needs to be extended to handle multiselection before this is fully
viable.

As part of this change, I needed to rewrite webtiles hover code to use a
regular css class rather than the :hover pseudo-class; unfortunately one
cannot interact programmatically with :hover. There *should* be no
user-visible change resulting from this but this was more complicated
than I had hoped, so we'll see. The webtiles client-server communication
about hover is still somewhat dubious here.