feat: hide dj MP bar in console/local tiles
This is kind of ugly because as it stands, y positions for certain hud
elements are calculated in multiple places. (In fact, this calculation
was already buggy.) But doing this "right" involves a lot of high-effort
refactoring and testing, or at least more than I'm going to do right
now. However, I have at least renamed the infamous yhack to something
more interpretable and changed its logic accordingly.