Title armour properties
To help clarify for newer players who get lost in the sea that
is our very long item descriptions. (We need more organization,
formatting, sections, etc... but instead I'm doing this.)

Example output:

'Of rampaging': It allows the wearer to take one free step toward
enemies. (Or however that goes...)