feat: quivering spells/potions
This is a pretty niche feature so I haven't included these in the
default `Q` list. However, it was easy to implement, requested, and this
is one step closer to a complete generic trigger api for item effects.
E.g. eventually I'd like to use actions for local tiles buttons so that
the item activation api is not distinct.

One noticeable gap is targeters for scrolls, to be added in a future
commit. Another is that scroll actions don't have a very good enabled
check.