refactor: generalize and unify some quiver boilerplate
This combines a bunch of shared / repeated code for actions involving
items. There's some hackiness here to do with hard coding class names
for serialization + equality purposes, but it's still probably better
than repeating this code so many times. (Not aware of a cleaner c++
approach to this, though perhaps there is something.)