feat: Make Lehudib's Crystal Spear cast quickly
Crawl has a very granular time system, but out side of weapon delay and
"what's the chance this monster double moves me" players don't have much
regular interaction with it (swap delays and that's it).

Crystal Spear is a cool spell. Big damage! Iconic name! Cool WizLab!
But, it's not so much better than Iron Shot that its cost is often
justified-between the shorter range, higher strategic and tactical cost
it's quite a bit.

One approach to making LCS more attractive is: bigger numbers! Yeah!
Bang! Pow! Except that, as the experience of A0 taught us, that might
get old if the accuracy also goes up. This commit takes a different
tack, making LCS "quick cast" when cast successfully, taking only half
the player's base delay (random rounded). A miscast takes the full time,
the time savings are flavoured as coming from the crystal!

Player-monster symmetry is a lie but in this case monsters also get the
technology. (Always rounding up because the divisor is buried deep in
the energy code and the rule is round_up and I'm not going to touch
that.) Casting frequency and spell selection AI means they won't turn
into machine gunners, but they will often get a crystal spear plus
something else if they decide to crystal spear first.