feat: New Spell: Maxwell's Coupling Level 8 Air magic
Commit 64d35965e29 removed Absolute Zero because the memorable,
spectacular instant kill effect didn't have good playfeel. It often felt
cheap, with the monster getting no chance to react. The result was that
boss fights felt like anticlimaxes.

This re-introduces the guaranteed kill effect, but it is no longer
instant.  Instead, the spell has a (random, not revealed) aut charging
time between cast and effect. In this time the player must wait (like
searing ray). At the end of the time a random closest monster is killed.
Any action other than waiting causes the spell to fizzle (with no MP
refund).

The flavour is currently Air: an oblique reference to capacitive
coupling. The effect (especially with the time lag) fits the "bouncy,
pointy, or hard to direct" archetype that we went for with Air in
positional magic much better than Ice (I will admit a bit of contortion
in calling A0 "subtle".) With Chain Lightning moved up to L9 Conj/Air,
Air magic has space for a L7 or 8 attack spell.

This reverts commit 64d35965e295382a687bd14c17aadd49688b771d.