feat: add static targeters for several scrolls
Another niche feature, somewhat more involved than I had hoped. The
trick here is that even where scrolls correspond to spells, the call
that does the effect is always pretty low-level and so it wasn't easy to
reuse all the your_spells scaffolding.

This commit adds targeters for fear, summoning, vulnerability,
immolation, holy word, silence, and reuses the existing blink targeting
code. (In implementing this, I slightly refactored how hop is called
because of needing to change shared code.) Most of these aren't very
exciting unless you aren't used to the effect, but fear seems useful. It
would be nice if there were a nicer interface for getting to these
targeters besides quivering the scroll and using `f`, but I'm not sure
what that would be.

This commit also prevents quivering unid'd spells and scrolls, for the
sake of simplicity.