feat: range checks for various scrolls
The main goal of this was to improve the enabled check when a scroll is
quivered (e.g. fear is now disabled with no susceptible enemies), but I
also added a double-check prompt when reading the scroll. If this is too
annoying it can be removed.

It's possible these checks are a bit slow when a scroll is quivered, and
need a cache mechanism like `spell_action`s have. However, I'm not sure
how much consumables will really be quivered in practice, so I'll wait
and see.